Move under `x` subdirectory
